Cambios en comentarios y usuarios

+++ IntroduccionEFCore/Entidades/Comentario.cs
@@ -6,7 +6,7 @@ namespace ServicaDB.Entidades
     {
         public int Id { get; set; }
         public string? Contenido { get; set; } = null!;
-        public int Calificacion { get; set; }
+        public decimal Calificacion { get; set; }
         public DateTime FechaCrea { get; set; }
         public DateTime FechaAnula { get; set; }
         public int PrestadorId { get;set; }//POR CONVENCION AL TENER LA PALABRA Id LUEGO DE LA ENTIDAD Prestador LO TOMA COMO CLAVE FORÁNEA
